[Refactor] #11 전체 패키지 구조 수정 (#12)
* rename: types 패키지 정리

각 관련 도메인 패키지 내부로 이동

* rename: 패키지 구조 전체 수정

* rename: entity 패키지 추가

* rename: 패키지 구조 변경
