+++ src/main/java/com/foodielog/server/admin/type/WithDrawReason.java
@@ -1,0 +1,5 @@
+package com.foodielog.server.admin.type;
+
+public enum WithDrawReason {
+
+}
