--- src/main/java/com/foodielog/server/feed/FeedLike.java
@@ -1,36 +1,0 @@
-package com.foodielog.server.feed;
-
-import com.foodielog.server.notification.Notification;
-import com.foodielog.server.user.User;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.CreationTimestamp;
-
-import javax.persistence.*;
-import java.sql.Timestamp;
-
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@Table(name = "feed_like_tb")
-@Entity
-public class FeedLike {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "feed_id")
-    private Feed feed;
-
-    @ManyToOne
-    @JoinColumn(name = "user_id")
-    private User user;
-
-    @ManyToOne
-    @JoinColumn(name = "notification")
-    private Notification notification;
-
-    @CreationTimestamp
-    private Timestamp createAt;
-}
