--- src/main/java/com/foodielog/server/reply/Reply.java
@@ -1,43 +1,0 @@
-package com.foodielog.server.reply;
-
-import com.foodielog.server.feed.Feed;
-import com.foodielog.server.types.ContentStatus;
-import com.foodielog.server.user.User;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.UpdateTimestamp;
-
-import javax.persistence.*;
-import java.sql.Timestamp;
-
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@Table(name = "reply_tb")
-@Entity
-public class Reply {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "feed_id")
-    private Feed feed;
-
-    @Column(length = 150, nullable = false)
-    private String content;
-
-    @Enumerated(EnumType.STRING)
-    private ContentStatus status;
-
-    @CreationTimestamp
-    private Timestamp createdAt;
-
-    @UpdateTimestamp
-    private Timestamp updatedAt;
-}
