--- src/main/java/com/foodielog/server/restaurant/RestaurantLike.java
@@ -1,31 +1,0 @@
-package com.foodielog.server.restaurant;
-
-import com.foodielog.server.user.User;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.CreationTimestamp;
-
-import javax.persistence.*;
-import java.sql.Timestamp;
-
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@Table(name = "restaurant_like_tb")
-@Entity
-public class RestaurantLike {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @ManyToOne
-    @JoinColumn(name = "restaurant_id")
-    private Restaurant restaurant;
-
-    @ManyToOne
-    @JoinColumn(name = "user_id")
-    private User user;
-
-    @CreationTimestamp
-    private Timestamp createdAt;
-}
