--- src/main/java/com/foodielog/server/types/BlockReason.java
@@ -1,4 +1,0 @@
-package com.foodielog.server.types;
-
-public enum BlockReason {
-}
