--- src/main/java/com/foodielog/server/types/ReportReason.java
@@ -1,10 +1,0 @@
-package com.foodielog.server.types;
-
-public enum ReportReason {
-    광고,
-    욕설,
-    명예훼손,
-    음란,
-    기타
-    ;
-}
