--- src/main/java/com/foodielog/server/types/WithDrawReason.java
@@ -1,5 +1,0 @@
-package com.foodielog.server.types;
-
-public enum WithDrawReason {
-
-}
