--- src/main/java/com/foodielog/server/user/WithDrawUser.java
@@ -1,33 +1,0 @@
-package com.foodielog.server.user;
-
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.CreationTimestamp;
-import org.hibernate.annotations.DynamicInsert;
-
-import javax.persistence.*;
-import java.sql.Timestamp;
-
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Getter
-@Table(name = "withdraw_user_tb")
-@Entity
-public class WithDrawUser {
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private User user;
-
-    @Column(name = "feed_count")
-    private Long feedCount;
-
-    @Column(name = "reply_count")
-    private Long replyCount;
-
-    @CreationTimestamp
-    private Timestamp createdAt;
-}
