+++ src/main/java/com/foodielog/server/user/type/Flag.java
@@ -1,0 +1,6 @@
+package com.foodielog.server.user.type;
+
+public enum Flag {
+	YES,
+	NO;
+}
